fix: Solve GAIA_PARTITION_SIZE var type issue

+++ code/scripts/downloader.py
@@ -5,6 +5,7 @@ from cdalvaro import DB, Logger
 from cdalvaro.catalogues import OpenClust
 from cdalvaro.downloaders import Gaia
 import logging
+import numpy as np
 import os
 
 parser = argparse.ArgumentParser(prog='Gaia DR2 Downloader', description='✨ Stars downloader from the GAIA DR2 dataset')
@@ -92,7 +93,7 @@ if len(clusters) == 0:
 gaia_username = os.getenv('GAIA_USER', None)
 gaia_password = os.getenv('GAIA_PASS', None)
 
-Gaia.partition_size = os.getenv('GAIA_PARTITION_SIZE', 500_000)
+Gaia.partition_size = np.int32(os.getenv('GAIA_PARTITION_SIZE', 500_000))
 
 gaia = Gaia(db=db, username=gaia_username, password=gaia_password)
 gaia.download_and_save(regions=clusters, extra_size=args.extra_size)
